fix two keys rotation logic to prevent downtime (#23)
fix two keys rotation logic to prevent downtimeFor a source with two (or multiple) keys, we should also check that the second newer key is older than the `max_age` before rotating the older key. This is so that we ensure that all jobs running on the older key have completed (e.g. if a job was started right before the newer key was created, and we run rotator again immediately after creating the newer key, then we don't want to delete the older key yet or the job would fail).

### Test Plan
unittests